MDEV-27602: Automatic creation of working directories in mtr
Our server has many settings for different directory paths:
--datadir=<path>, --log-bin=<path>, --innodb-data-home-dir=<path>
and so on. But we do not have a sufficient number of tests that
would check the operation of these settings in combination with
such complex subsystems as, for example, Galera + mariabackup
(for SST).

Creating such tests can be a non-trivial task. For example,
for a Galera cluster, to change the path for --innodb-data-home-dir,
we need to stop the first node, then create a new directory structure
for the database, restart the first node - but not as the main node
of the cluster, but as a joiner, then repeat the same manipulations
for the second node. Then, after the test, we must destroy the source
files and restart both nodes in sequence with default settings,
otherwise mtr will detect a configuration discrepancy or subsequent
tests will fail. And after that, we must carefully remove all the
"junk" directories that were created during the test.

This not only generates dozens of lines of code in .test files
that need to be maintained when changes are made (for example, if
new database state files or new types of logs are added to innodb),
but also a source of problems when restarting tests.

If such a test fails, its "junk" directory structure is not
completely destroyed. If some directory with the same name is
used in another similar test, this will lead to false failures.
Similarly, when restarting tests. And mtr does not have enough
facilities to detect and destroy such temporary working structures
before starting the test - unless we use system-specific inserts
in the form of .sh scripts, --exec or perl inserts.

Also, mtr does a good job of bootstrap the initial base from
the initial state (from the install.db directory), however if
we are testing a cluster (for example Galera) with non-standard
directories, then we have to actually redo all the work that mtr
does for bootstrap, plus with manually restarting nodes to change
directories.

Therefore, this commit introduces a change to mtr that allows us to
automatically create temporary directories before starting individual
tests and automatically destroy them after the end of the test,
including if the test fails. Also, if such temporary directories
are left over from previous failed test runs, but they are reused
in other tests, they will be automatically destroyed and recreated
before starting a new test.

For each test, a ".dirs" file can now be created that lists the
directories that must be created before starting the test. Then mtr
will automatically create these directories before starting the first
server (if necessary, destroying previous versions left over from
failed previous tests). After the test completes (whether normal or
due to failure), this directory will be automatically deleted by mtr
itself before running the next tests. And all test servers will be
automatically restarted again.

In this case, paths to automatically created directories can use
variable substitution through the standard ${variable_name} construct,
for example, like this: ${MYSQLTEST_VARDIR}/mysqld.1/innodb_data

If the paths to these temporary directories are specified in .cnf
files or in .opt files, then the bootstrap procedure will immediately
create all the database files in the working directories.

For example, this allows you to create tests that check work when
changing paths to some innodb working directories by copying code from
regular tests (which work with standard paths) - without interfering
with the structure of the .test file, only by changing the settings in
the .cnf file copied from regular test. Say, to test the operation of
the Galera cluster with SST through mariabackup in combination with
--innodb-data-home-dir=<non-default-path>, it is enough to copy the
existing test (for example, "galera_sst_mariabackup.test"), add a
".dirs" file for it with two lines (for two nodes):

${MYSQLTEST_VARDIR}/mysqld.1/innodb_data
${MYSQLTEST_VARDIR}/mysqld.2/innodb_data

and then we need to write these paths to .cnf file or .opt file.
And automatically we get a test that will check the operation of the
cluster with a non-standard data directory for innodb. That being said,
we don't need to manually write code to restart the cluster with the
new settings in the .test file, we don't need to repeat the bootstrap
work there, and we don't need to worry about destroying those working
directories at the end of the test.

In addition, lines in the .dirs file may contain an "=" sign after
the name of the temporary directory, followed by the path to another
already existing directory. In this case, mtr will itself copy the
entire subtree of files from this directory to the working directory
(and destroy everything after the test is completed). This will make
it easier to create tests that take some reference files (say from
std_data) and then modify them during the test.

+++ mysql-test/mysql-test-run.pl
@@ -2773,8 +2773,8 @@ sub mysql_server_start($) {
     {
       # Some InnoDB options are incompatible with the default bootstrap.
       # If they are used, re-bootstrap
-      if ( $extra_opts and
-           "@$extra_opts" =~ /--innodb[-_](?:page[-_]size|checksum[-_]algorithm|undo[-_]tablespaces|log[-_]group[-_]home[-_]dir|data[-_]home[-_]dir)|data[-_]file[-_]path/ )
+      if ( $tinfo->{dirs} or ($extra_opts and
+           "@$extra_opts" =~ /--innodb[-_](?:page[-_]size|checksum[-_]algorithm|undo[-_]directory|undo[-_]tablespaces|log[-_]group[-_]home[-_]dir|data[-_]home[-_]dir|buffer[-_]pool[-_]filename)|data[-_]file[-_]path/) )
       {
         mysql_install_db($mysqld, undef, $extra_opts);
       }
@@ -3771,6 +3771,107 @@ sub resfile_report_test ($) {
   resfile_test_info("start_time", isotime time);
 }
 
+sub create_dirs ($) {
+  my $tinfo = shift;
+  my $res = 0;
+  if ($tinfo->{dirs}) {
+    open my $file, $tinfo->{dirs} or $res = 1;
+    if ($res == 0) {
+      while (my $line = <$file>) {
+        chomp $line;
+        $line =~ s/\$\{([^}:]+)(:([^}]+))?\}/defined $ENV{$1} ? $ENV{$1} : $3/eg;
+        $line =~ s/^\s+|\s+$//g;
+        my $dir = $line;
+        my $src = "";
+        if (index($dir, '=') != -1) {
+          ($dir, $src) = $dir =~ /^([^=\s]*(?:\s+[^=\s])*)\s*=\s*(.*)$/;
+        }
+        if ($dir eq "") {
+           next;
+        }
+        if (!($dir =~ /^[\/\\][^\/\\]*$/)) {
+          if (-d $dir) {
+            rmtree($dir);
+            if (-d $dir) {
+              $res = 1;
+              last;
+            }
+          }
+          if ($src eq "") {
+            mkpath($dir);
+          }
+          else {
+            copytree($dir, $src);
+          }
+          if (! -d $dir) {
+            $res = 1;
+            last;
+          }
+        }
+        else {
+          mtr_print("Operations with the root of FS are prohibited");
+          $res = 1;
+          last;
+        }
+      }
+      close $file;
+    }
+    # Create temporary directories:
+    if ($res != 0) {
+      my $diag= "Failed to create directories specified in '$tinfo->{dirs}'";
+      $tinfo->{'comment'} = $diag;
+    }
+  }
+  return $res == 0;
+}
+
+sub delete_dirs ($) {
+  my $tinfo = shift;
+  my $res = 0;
+  if ($tinfo->{dirs}) {
+    open my $file, $tinfo->{dirs} or $res = 1;
+    if ($res == 0) {
+      while (my $line = <$file>) {
+        chomp $line;
+        $line =~ s/\$\{([^}:]+)(:([^}]+))?\}/defined $ENV{$1} ? $ENV{$1} : $3/eg;
+        $line =~ s/^\s+|\s+$//g;
+        my $dir = $line;
+        if (index($dir, '=') != -1) {
+          ($dir) = $dir =~ /^([^=\s]*(?:\s+[^=\s])*)\s*=.*$/;
+        }
+        if ($dir eq "") {
+           next;
+        }
+        if (!($dir =~ /^[\/\\][^\/\\]*$/)) {
+          if (-d $dir) {
+            rmtree($dir);
+            if (-d $dir) {
+              $res = 1;
+              last;
+            }
+          }
+        }
+        else {
+          mtr_print("Operations with the root of FS are prohibited");
+          $res = 1;
+          last;
+        }
+      }
+      close $file;
+    }
+    # Delete temporary directories:
+    if ($res != 0) {
+      my $diag= "Failed to delete directories specified in '$tinfo->{dirs}'";
+      if ($tinfo->{'comment'}) {
+        $tinfo->{'comment'} .= "\n".$diag;
+      }
+      else {
+        $tinfo->{'comment'} = $diag;
+      }
+    }
+  }
+  return $res == 0;
+}
 
 #
 # Run a single test case
@@ -3903,7 +4004,7 @@ sub run_testcase ($$) {
       My::SafeProcess->start_exit();
     }
 
-    if (start_servers($tinfo))
+    if (create_dirs($tinfo) && start_servers($tinfo))
     {
       report_failure_and_restart($tinfo);
       unlink $path_current_testlog;
@@ -4064,6 +4165,10 @@ sub run_testcase ($$) {
           }
         }
         mtr_report_test_passed($tinfo);
+        if ($tinfo->{dirs}) {
+          stop_all_servers($opt_shutdown_timeout);
+          delete_dirs($tinfo);
+        }
       }
       elsif ( $res == 62 )
       {
@@ -4074,9 +4179,11 @@ sub run_testcase ($$) {
         mtr_report_test_skipped($tinfo);
         # Restart if skipped due to missing perl, it may have had side effects
 	if ( restart_forced_by_test('force_restart_if_skipped') ||
-             $tinfo->{'comment'} =~ /^perl not found/ )
+             $tinfo->{'comment'} =~ /^perl not found/ ||
+             $tinfo->{'dirs'} )
         {
           stop_all_servers($opt_shutdown_timeout);
+          delete_dirs($tinfo);
         }
       }
       elsif ( $res == 65 )
@@ -4879,7 +4986,6 @@ sub report_failure_and_restart ($) {
   my $test_failures= $tinfo->{'failures'} || 0;
   $tinfo->{'failures'}=  $test_failures + 1;
 
-
   if ( $tinfo->{comment} )
   {
     # The test failure has been detected by mysql-test-run.pl
@@ -4924,6 +5030,8 @@ sub report_failure_and_restart ($) {
   }
 
   after_failure($tinfo);
+
+  delete_dirs($tinfo);
 
   mtr_report_test($tinfo);
 
@@ -5182,9 +5290,15 @@ sub server_need_restart {
     return 1;
   }
 
-  if ( $tinfo->{'master_sh'}  || $tinfo->{'slave_sh'} )
+  if ( $tinfo->{'master_sh'} || $tinfo->{'slave_sh'} )
   {
     mtr_verbose_restart($server, "sh script to run");
+    return 1;
+  }
+
+  if ( $tinfo->{'dirs'} )
+  {
+    mtr_verbose_restart($server, "uses additional directories");
     return 1;
   }
